refactor: upgrade eslint
