+++ .github/pull-request-template.md
@@ -3,6 +3,6 @@
 
 1.  fix bug, create item and remove.
 
-### I running lint, tsc, tests and commit lint?
+### I running lint, tsc, tests and commit lint
 
 -   [] yes
