+++ package.json
@@ -62,6 +62,7 @@
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-promise": "^6.0.0",
     "eslint-plugin-sonarjs": "^0.15.0",
+    "eslint-plugin-spellcheck": "^0.0.20",
     "husky": "^8.0.1",
     "jest": "^29.0.1",
     "lint-staged": "^13.0.4",
