+++ tsconfig.json
@@ -7,6 +7,7 @@
     "strict": true,
     "noImplicitAny": true,
     "strictNullChecks": true,
+    "noErrorTruncation": true,
     "outDir": "./dist",
     "rootDir": "./src",
     "removeComments": true,
